requirements.txt
certifi==2024.2.2
emoji==2.10.1
feedparser==6.0.11
future==0.18.3
python-dateutil==2.8.2
python-telegram-bot==20.7
pytz==2024.1
six==1.16.0
peewee==3.17.0